Create singleton nvmemallocator
